:recycle: move crawler to separate folder
