+++ flathunter/config.py
@@ -9,13 +9,13 @@ from dotenv import load_dotenv
 from flathunter.captcha.captcha_solver import CaptchaSolver
 from flathunter.captcha.imagetyperz_solver import ImageTyperzSolver
 from flathunter.captcha.twocaptcha_solver import TwoCaptchaSolver
-from flathunter.crawl_ebaykleinanzeigen import CrawlEbayKleinanzeigen
-from flathunter.crawl_idealista import CrawlIdealista
-from flathunter.crawl_immobiliare import CrawlImmobiliare
-from flathunter.crawl_immobilienscout import CrawlImmobilienscout
-from flathunter.crawl_immowelt import CrawlImmowelt
-from flathunter.crawl_wggesucht import CrawlWgGesucht
-from flathunter.crawler_subito import CrawlSubito
+from flathunter.crawler.ebaykleinanzeigen import CrawlEbayKleinanzeigen
+from flathunter.crawler.idealista import CrawlIdealista
+from flathunter.crawler.immobiliare import CrawlImmobiliare
+from flathunter.crawler.immobilienscout import CrawlImmobilienscout
+from flathunter.crawler.immowelt import CrawlImmowelt
+from flathunter.crawler.wggesucht import CrawlWgGesucht
+from flathunter.crawler.subito import CrawlSubito
 from flathunter.filter import Filter
 from flathunter.logging import logger
 from flathunter.exceptions import ConfigException
